Rewrite code to check for bubbles to drop
Wasn't taking into account bubbles that are not neighbors of the popped.
Now have a function to get all connected bubbles to a given bubble and
use that to find all connected to ceiling bubbles.
Go over to_check_drop and if not in connected,
drop that one and all connected to it.